Remove plane_datapoint
