+++ lib/view/data/index.ts
@@ -1,5 +1,4 @@
 export * from './chart_landing';
 export * from './data';
 export * from './datapoint';
-export * from './plane_datapoint';
 export * from './series';
